Add query UUIDs to read command
This patch adds three parameters to read_command which are needed for
caching of readers during multiple pages of a paged queries, which
we will introduce in the next patch.

The reader_recall_uuid is a UUID of a previously saved reader, which
the replica is now asked to recall and resume (if this saved reader is
no longer in the cache, it is fine, a new reader will be started).

The reader_save_uuid is the UUID to save the reader after reading the
current page of data. We deliberately want to change the UUID under
which we save the reader, so that a repeat of the same query (e.g.,
the client resends a query after the coordinator timed out) will
not accidentally re-use a read which has already been read from and
now points at a different position.

The reader_recall_drop flag allows recalling a saved reader but
immediately dropping it instead of reusing it. This will be needed
in some cases (namely, reconciliation) where we know the reader
we previously saved is not in the correct position and should be
dropped immediately.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>