In paged queries, resume readers instead of creating new ones
As currently implemented, each of the individual page requests opens
a new reader, with all the overhead this entails - finding the right
sstables, reading from their index files, and reading large buffers (and
even more with read-ahead) from each sstable. Most of this work is lost and
needs to be redone for the next page.

The goal of this patch is to save the reader when finishing one page, and
continuing the same reader (without opening a new one) when the following
page is requested.

Saving a reader is not as simple as it sounds: The existing code destroys
the mutation_reader when done with it, so we need to wrap it in an wrapper
which will be destroyed but the instead of destroying the reader it contains,
it saves it in a cache. Also read may (and in general, always will) stop
inside a partition, and we also need to save the current streamed_mutation
state instead of destroying it. There are multiple complications in doing
this, as you can see in the code (and comments).

Fixes #1865.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>