Fix failure to parse the type name "list<empty>" in schema table
In schema_tables.cc, where we persist schemas to disk, we write the types of
columns to disk as CQL-like strings using type->as_cql3_type()->to_string().
We later read this string and convert it back to a type with the function
db::cql_type_parser::parse().

The parse() functions is supposed to recognize every type name we could
concievably write. But because the way it is currently written, it misses
some cases, such as "list<empty>" (and similarly for other collection types
containing items of the type "empty"):

parse() starts by checking if the type is one of the native types, but if
it isn't, it falls back to using CQL's type parser. So we have a problem
with a type like empty_type which has a name ("empty") but is not supported
by CQL syntax: Before this patch, parse() recognized "empty" (which is in
the list of native types) but could not parse "list<empty>", because CQL
doesn't recognize "empty" as a valid type name.

This problem did not cause problems in existing code, because our only use
of empty_type has been as an empty column added to dense tables without regular
columns. We never had a list<empty> column anywhere (and users couldn't add
one because there was no CQL syntax for it). But in the following patches
for fixing #3362, we will start to add collections of empty as well, and
we will need this case to be parsed correctly.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>