Do not allow selecting a virtual column
For issue #3362, we will need to add to a materialized view also unselected
base-table columns as "virtual columns". We need these columns to exist
to keep view rows alive, but we don't want the user to be able to see
them.

In this patch we prevent SELECTing the virtual columns of the view,
and exclude the virtual columns from a "SELECT *" as well.

Refs #3362.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>