alternator: reimplement read-modify-write operations using LWT
In this patch, we re-implement the three read-modify-write operations -
PutItem, UpdateItem, DeleteItem. All three operations may need to read the
item before writing it to support conditional updates (the "Expected"
parameter) and UpdateItem may also need the previous item's value for
its update expression (e.g., a user may ask to "set a=a+1" or "set a=b").

Before this patch, the implementation of RMW operations simply did a read,
and then a write - without any attempt to protect concurrent operations.

In this patch, Scylla's LWT mechanism (storage_proxy::cas()) is used
instead, to ensure that concurrent update operations are correctly
isolated even if they are conditional.

The version presented here is configured to always use LWT for *every*
write, regardless of whether it has a condition or not. This is why I
marked this version RFC - it will significantly slow down even write-only
workloads like YCSB, and we should measure and consider whether this is
acceptable.
But the code in this patch actually includes other modes, which can be
chosen by setting a boolean constant in the code. In the future we may
want (?) to let the user configure these flags, globally or per table or
per attribute:

1. The "forbid_rmw" flag can be set to forbid any RMW operation.

2. The "non_rmw_uses_lwt" can be turned off so that non-RMW writes
   will do a regular write instead of LWT. This is safe if "forbid_rmw"
   is on, or if you're sure that the LWT and non-LWT go to different
   tables, items or even attributes.

3. The "unsafe_read_before_write" flag returns the old behaviour,
   of separate read and then write. This mode is not recommended, and
   we'll probably drop it after completing the testing.

Note that in this implementation (and, hopefully, this decision will remain),
read requests are NOT modified, and work exactly as they did before: i.e.,
strongly-consistent reads are done using a normal CL=LOCAL_QUORUM read -
not via LWT. I believe this is correct "enough", and that anything else we
will do will ruin our performance.

Unfortunately, this patch involves a large amount of code reorganization,
because:
1. The cas operation requires each operation to be made into an object,
   with a separate apply() function, forcing a lot of code to move.
2. Moreover, we need to do this for three different operations (PutItem,
   UpdateItem, DeleteItem) so to avoid massive code duplication, I had
   to move some common code.
3. The cas operation also forced us to change some of the utility functions'
   APIs.

The end result is that this patch focuses more on a compact and
understandable *end result* than it does on an easy to understand *patch*,
so sorry about that.

All alternator-test/ tests pass with this patch (and also with all of the
different optional modes enabled). However, other than that, I did not yet
do any real isolation tests (are concurrent operations really isolated
correctly? or is LWT just faking it? :-) ), performance tests or stress
tests - and I'll definitely need to do those as well.

Fixes #5054

Signed-off-by: Nadav Har'El <nyh@scylladb.com>