alternator: add example DNS load balancer implementation
This is a proof-of-concept implementation for a DNS-based load balancing
technique for Alternator, as described in the "The API Server" section of
https://docs.google.com/document/d/1i4yjF5OSAazAY_-T8CBce9-2ykW4twx_E_Nt2zDoOVs

Note that DNS-based load balancing has upsides and downsides. The upside is
that it requires very little resources. Ideadlly an existing DNS server could
be reused (though this is not demonstrated here). The main downside is that
it is not very useful when the number of different client machines is low,
since each client machine will likely contact the same server node for the
entire TTL period (4 seconds).

In any case, this is a demonstration of easy it is to do this sort of load
balancing in just 50 lines of code. The example code starts with one known
node in the cluster (127.0.0.1), and then asks it - via the request
http://127.0.0.1/localnodes, for an updated list of nodes in this DC; This
is repeated once every second, and each time a user asks for name resolution
one of the nodes in the most recent list is returned at random.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>