alternator-test: a "run" script
Running the Alternator tests is easy after you manually run Scylla, but
sometimes it's convenient to have a script which just does everything
automatically: start Scylla in a temporary directory, set it up properly
for the tests (namely, the authentication), run all the tests, and clean
up the temporary directory. This is what this script does.

This scripts can be run by Jenkins, for example, to check all the Alternator
tests. The script assumes some things (including cqlsh, pytest and the boto3
library) are already installed, and Scylla is compiled in build/dev/scylla
but this can be overridden by a command like

    SCYLLA=build/release/scylla alternator-test/run

Signed-off-by: Nadav Har'El <nyh@scylladb.com>