build: switch "dev" build from -O1 to -O2
Following the switch of Scylla to clang, we notice a big slowdown of code
(Scylla and tests) compiled in "dev" build mode.

In this patch I propose changing the dev build from -O1 to -O2.
The consequences of this change will be:

1. A small slowdown of the build.
   I measured building build/dev/scylla across two build machines (with
   distcc) and without ccache. Before this patch, the build took 798
   seconds, after this patch, it took 872 seconds. That's a 9% slowdown.

2. A huge speedup of the resulting code.
   As an example, I measured "test/alternator/run". Before this patch, it
   needed 3 seconds to boot Scylla and 118 seconds to complete the tests.
   After this patch, the numbers are down to 1 second and 48 seconds!
   That's a 2.5-fold speedup.

I admit that this is a tradoff, but I think it's a good tradeoff.
A 9% build slowdown is hardly noticable, but a 2.5-fold slowdown
of tests will cause developers to think twice before running tests,
which is bad.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>