cql-pytest: reproduce two more bugs in toJson()
Add two xfailing tests reproducing two issues - #7988 and #8002.
The first is that toJson() incorrectly formats values of the "time"
type - it should be a string but Scylla forgets the quotes.
The second is that toJson() format "decimal" values as JSON numbers
without using an exponent, resulting in memory allocation failure
for numbers with high exponents, like 1e1000000000.

Both tests pass on Cassandra, and fail on Scylla.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>