alternator: support attribute paths in ConditionExpression
This patch fully implements support for attribute paths (e.g. a.b.c, a.d[3])
for the ConditionExpression in conditional updates. After this patch, the
two xfailing tests in test_projection_expression.py now pass.

The fix is simple: When calculating the value of an expression in
calculate_value(), when we want to evaluate a path, we can't just take
the top-level attribute - we need to walk into the specific sub-object
as specified by the attribute path.

This is not the end of attribute path support, there are still other
expressions (UpdateExpression, FilterExpression, ReturnValues) where
they are not yet supported. This will come in following patches.

Refs #5024

Signed-off-by: Nadav Har'El <nyh@scylladb.com>