test/boost: add unit test for Heat-Weighted Load Balancing
In issue #9053 a problem in the test was theorized to result from a bug
in HWLB (Heat-Weighted Load Balancing) which may send too many requests
to a replica with a low cache heat (i.e., low cache-hit ratio).

This patch contains a unit test attempting to prove or disprove this
theory. The unit test checks just the low-level function
miss_equalizing_combination() from db/heat_load_balance.{cc,hh} which
picks the random replicas to send to based on the known hit-ratios.
This function is the crux of the HWLB algorithm - its design is presented
in https://docs.google.com/document/d/1M-gRlPDPvqB8mWKBV1ix3AX29JCOX11zgVrtO2bf--E
These tests do not check some driving code in db/consistency_level.cc,
and do not use an actual Scylla cluster.

The results of these test brings good news, and bad news:

The good news is that in all combinations of number of replicas and
their hit ratios which I tried in the test, HWLB is tested to produce
the correct distribution of requests - leading to equal miss rates on
all replicas and running as many of the requests as allowed locally.

But the bad news is that the a few of the tests (invalid_combinations_*)
reproduce a bug described in issue #9285: Some of the combinations
of replicas returned by miss_equalizing_combination() are invalid -
listing the *same* replica more than once. This can cause CL violation -
a user asked for a read with CL=2 but the read was sent twice to the
same node, so really got just CL=1.

The failing tests are marked with "expected_failures" allowing us to
keep their code live but not fail the test suite.

In addition to those failing tests, it turns out that for some
succeeding tests, very rarely (about one in a million calls to
miss_equalizing_combination()) an invalid combination is returned.
I am guessing that this might be related to floating-point inaccuracy
leading to picking the same replica again instead of the next one.
Currently we work around this flakyness by allowing a small number
of invalid combinations (the failing tests mentioned above see
many thousands of invalid combinations).

The silver lining is that all known problems mentioned above involve RF>3.
I have not been able to reproduce any bug with RF=3, which is the most
common use case.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>