Alternator: drop unneeded "IS NOT NULL" clauses in MV of GSI/LSI
Scylla's materialized views naturally skip any base rows where the view's
key isn't set (is NULL), because we can't create a view row with a null
key. To make the user aware that this is happening, the user is required
to add "WHERE ... IS NOT NULL" for the view's key columns when defining
the view. However, the only place that these extra IS NOT NULL clauses
are checked are in the CQL "CREATE MATERIALIZED VIEWS" statement - they
are completely ignored in all other places in the code.

In particular, when we create a materialized view (GSI or LSI) in
Alternator, we don't have to add these "IS NOT NULL" clauses, as they
are outright ignored. We didn't know they were ignored, and made an
effort to add them - but no matter how incorrectly we did it, it didn't
matter :-) In commit 2bf2ffd3ed25b01527695262b462f14fe1422e53 it turned
out we had a typo that caused the wrong column name to be printed.
Also, even today we are still missing base key columns that aren't
listed as a view key in Alternator but still added as view clustering
keys in Scylla - and again the fact these were missing also didn't
matter. So I think it's time to stop pretending, and stop calculating
these "IS NOT NULL" strings, so this patch outright removes them from
the Alternator view-creation code.

Beyond being a nice cleanup of unnecessary and inaccurate code, it
will also be necessary when we add in later patches "computed columns",
where an Alternator attribute "x" isn't a real column x in the base
table but rather an element in the ":attrs" map - so adding a
"x IS NOT NULL" isn't only unnecessary, it is outright illegal:
The expression evaluation code, even though it doesn't do anything with
the "IS NOT NULL" expression, still verifies that "x" is a valid column,
which it isn't.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>