materialized views: pass row to compute_value()
We have in materialized views the notion of a "computed column", where
the view picks up not a real column from the base table but rather a
value somehow computed from the base table. The first computed column
was used to calculate the token of the partition key, so didn't need
anything except the partition key - so unfortunately this is all that
was passed to the virtual compute_value() function. But the caller
knows more - it also knows how the row is being updated, and at least
parts of existing row, so we should pass those too.

Unfortunately, when collection indexing was recently introduced, it
also needed the full row (and not just the partition key), but didn't
change compute_value() and instead added a new interface and added
update and existing to it. Now it's about time we change the original
interface as well - we want to write in a later patch a computed
column which extracts a single member out of a map in Alternator.

Beyond the code in view.cc which calls compute_value() with all the
added parameters, we also have additional secondary-index code which
*assumes* it only deals with the {legacy,}token_column_computation
computations it added, and doesn't have an update/existing to pass.
For this use, we also add a compute_value() overload that likes before,
only takes the partition key. Only {legacy,}token_column_computation
will implement this overload.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>